pyproject.toml
[project]
name = "dot-work"
version = "0.1.1"
description = "Portable AI coding prompts for project scaffolding and issue tracking"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["ai", "prompts", "scaffolding", "copilot", "claude", "cursor"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "radon>=6.0.0",
    # version module dependencies
    "GitPython>=3.1.0",
    "python-dotenv>=1.0.0",
    # zip module dependencies
    "gitignore-parser>=0.1.0",
    # review module dependencies
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.6.0",
    # container provision module dependencies
    "python-frontmatter>=1.1.0",
    # overview module dependencies
    "libcst>=1.1.0",
    # git module dependencies
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
# zip optional features (zip module upload)
zip-upload = ["requests>=2.28.0"]
# kg optional features (knowledge graph module)
kg-http = ["httpx>=0.27.0"]
kg-ann = ["hnswlib>=0.8.0"]
kg-all = [
    "httpx>=0.27.0",
    "hnswlib>=0.8.0",
]
# git module optional features (LLM summarization)
git-llm = [
    "openai>=1.0.0",
    "anthropic>=0.3.0",
]
# scan optional features (dependency visualization)
scan-graph = ["networkx>=3.0", "pyvis>=0.3.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
dot-work = "dot_work.cli:app"
kg = "dot_work.knowledge_graph.cli:app"
pybuilder = "dot_work.python.build.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dot_work"]
artifacts = [
    "src/dot_work/prompts/*",
    "src/dot_work/review/templates/*",
    "src/dot_work/review/static/*",
]

[tool.hatch.build.targets.sdist]
include = [
    "src/dot_work",
    "src/dot_work/prompts/*",
    "src/dot_work/review/templates/*",
    "src/dot_work/review/static/*",
]

[tool.uv]
package = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# S701: Autoescape disabled - intentional for markdown templates (not HTML)
"src/dot_work/installer.py" = ["S701"]
# S310: urllib.request calls - intentional HTTP API requests to embedding providers
"src/dot_work/knowledge_graph/embed/ollama.py" = ["S310"]
# B008, B904: Typer CLI patterns - function calls in defaults are standard Typer usage
"src/dot_work/container/provision/cli.py" = ["B008", "B904"]
"src/dot_work/python/build/cli.py" = ["B008", "B904"]
"src/dot_work/python/__init__.py" = ["B008", "B904"]
"src/dot_work/python/scan/cli.py" = ["B008", "B904"]
"src/dot_work/version/cli.py" = ["B008", "B904"]
"src/dot_work/overview/cli.py" = ["B008", "B904"]
# import-untyped: frontmatter library has no type stubs
"src/dot_work/container/provision/core.py" = ["S603"]
"src/dot_work/container/provision/validation.py" = []
"src/dot_work/knowledge_graph/embed/openai.py" = ["S310"]
# S112: Bare except-continue - intentional best-effort batch embedding storage
"src/dot_work/knowledge_graph/search_semantic.py" = ["S112"]

[tool.pytest.ini_options]
addopts = ["-v", "--tb=short", "--strict-markers"]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
